refactor Nav component: create ProfileMenu component
